python_hacking/seccion1/1_1_hacking_buscadores/requirements.txt
requests
python-dotenv
rich
gpt4all
openai
transformers